move procedures out of expressions
