+++ packages/cypher-builder/src/procedures/apoc/date.ts
@@ -17,10 +17,10 @@
  * limitations under the License.
  */
 
-import type { CypherEnvironment } from "../../../Environment";
-import { CypherASTNode } from "../../../CypherASTNode";
-import type { PropertyRef } from "../../../variables/PropertyRef";
-import type { Variable } from "../../../variables/Variable";
+import type { CypherEnvironment } from "../../Environment";
+import { CypherASTNode } from "../../CypherASTNode";
+import type { PropertyRef } from "../../variables/PropertyRef";
+import type { Variable } from "../../variables/Variable";
 
 export class ConvertFormat extends CypherASTNode {
     private temporalParam: Variable | PropertyRef;
